Customer productDetail view
first update, config router
